[#1] Factor out BaseSecretFormLite
* Moves the core secret form logic from SecretFormLite to a shared
  component for better reusability.
* Simplifies the Homepage component by removing
  redundant section wrapping.
